refactor(cockpit): extract change surface module
